// frontend/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BookResponse {
  archived?: boolean;
  authorName?: string;
  bookCover?: string;
  cover?: Array<string>;
  createdAt?: string;
  createdBy?: string;
  id?: string;
  isbn?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
  updatedAt?: string;
  updatedBy?: string;
}